[DM-27441] Use docker creds
This integrates the usage of the docker credential as a kubernetes
secret.

First, we have the name of the secret passed in as an environment
variable, which we use to pass it into the daemonset.

It's also mounted so we can use the same credential store to lookup
the username and password to ask the registry for the tag listing.