[DM-27441] Fix my labeling
So kubernetes does labels as a dict of string:string.  Fair enough.
I wasn't doing it that way, but that seems much better.  So change
over to using a set of labels (which can always be one).  What I'm
not handling is the logic part, I'm just doing simple subset of
labels.