[DM-27441] Hold preliminary results in local variables
So, this shouldn't have been able to be interrupted to give a
result which isn't consistent.  It should be either what it
was previously, or the new result, depending on when the request
is scheduled vs the do_work, and since python runs on 1 thread,
this shouldn't be a problem.

But if we use a threadpool or change things, it might not keep,
so I'm going to build up in local variables and then assign it
to the instance variables at the end.  We may need to add a lock
around those three lines if we run on different threads.  For
now, since a look isn't absolutely required, and it adds confusion,
I'm going to leave it without a lock.