[DM-27441] Clean up tox tests
First, up black which is run by tox to 3.8.

Next, move kubeconfig.load_incluster_config into the create_app
instead of import at module time.  When pytest imports this, it
does it before the mocks are applied, so it wouldn't call the mock.
Better to not call things on module import anyway.

Get rid of some of the extra test directories and make things simpler,
since there is no internal or external stuff.  It's all external.