[DM-27441] Redo the recommended tag
Rejigger this work on the recommended tag.  Instead of trying
to filter down the cache to what we were looking for, just pass
the whole thing into the repoman interface.  That also makes the
repoman interface simpler too.  This also allows us to make sure
the hash is the same as the one in the registry, for times where
we are repushing the same tag, like recommended.