[DM-27441] Implement docker registry for aiohttp
First, get rid of DXF dependency, which uses requests, and was
trouble mixing async and sync together in an aiohttp server app.

Then, take what's in dockercreds and make it into a full client
that can get the tags and digests.  Change the calls to make it
await, and get the tests running again.