[DM-27441] Recommended_image_url -> recommended_tag
I used to have it be tag, then changed it to the URL, but I think
it's easier to work with as the tag.  Also fix up the tests a bit,
since I was passing in the URL (which wasn't used, but was accepted
by the schema checking), but for the standalone rubin tests, it was
the tag only.