[DM-28601] Always use full image names
Since the names used by the images can seemingly be either short
or long, and depend it seems on how they are first pulled, just
try to use the long name everywhere.  Fix the tests to reflect it.

Also, if the hash matches and there is one, just say we have it
even if the image url doesn't exactly match - it's the same hash.