[DM-28601] More thorough cache checking
So docker caches from kubernetes point of view are organized by
the hashes, and not by the names or where they are pulled.  So if
you retagged an image and pushed it, it would show up next to it.

So first, undo what I did before, which loosened the restrictions
but wasn't exactly perfect.  Make the tests better to add the images
to look the same.  Add a test to the pull test that pretends to be
the prepuller to add things to the cache.  Then make sure everything
is fine.