[DM-28601] Loosen up hash checks
So, sometimes apparently there is a set of images that don't have
a hash coming out of kubernetes.  At least on Google.  I'm not
sure why this doesn't have a hash, maybe it's not pushed somewhere
or something.

Here's the entry of the troublesome freeloader:

    - names:
      - gke.gcr.io/kube-proxy-amd64:v1.18.12-gke.1205
      - k8s.gcr.io/kube-proxy-amd64:v1.18.12-gke.1205
      sizeBytes: 120890453

This is gotten by running kubectl get nodes -o yaml