[DM-28726] Delete dangling daemonset
So with the order of these in the if statement, we would end up
leaving the last daemonset running forever.  This wasn't really
a problem in pulling it (we find it after the first iteration),
but it's much nicer to know that only things that are working
are running.  So we should check if we're caching first, which
will delete anything that is dangling.