[DM-30180] Add owner references to created pods
This allows for argocd to know that the pods are linked
back to the cachemachine pod.  This also works at the
kubernetes layer, so that if you delete the deployment
for cachemachine, it should delete the pods, which deletes
the daemonset, and then the pulling pods.

This might be a bit much, but it's a safer sounding upgrade
strategy.